pyproject.toml
[project]
name = "crawfishbot"
version = "0.1.0"
description = "crawfishcomic Wiki helper bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["pywikibot", "pytumblr", "mwparserfromhell", "requests", "pytz", "python-dateutil", "feedparser"]
